etl_transformer/requirements.txt
requests>=2.25.0
psycopg2-binary>=2.8